% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggphylomorpho.R
\name{ggphylomorpho}
\alias{ggphylomorpho}
\title{Make a phylomorphospace plot}
\usage{
ggphylomorpho(tree, tipinfo, xvar = PC1, yvar = PC2, factorvar = group,
  labelvar = taxon, title = "Phylomorphospace", xlab = "PC1",
  ylab = "PC2", repel = TRUE, edge.width = 1, fontface = "italic",
  tree.alpha = 0.7)
}
\arguments{
\item{tree}{object of class \code{phylo}}

\item{tipinfo}{a dataframe containing two continuous variables (typically PC1 and PC2 from a PCA) as well as a grouping variable and a variable of labels for each point on the plot}

\item{xvar}{the unquoted name of the column containing the x variable in the tipinfo dataframe. Defaults to \code{PC1}}

\item{yvar}{the unquoted name of the column containing the y variable in the tipinfo dataframe. Defaults to \code{PC2}}

\item{factorvar}{the unquoted name of the column containing the grouping variable in the tipinfo dataframe. Defaults to \code{group}}

\item{labelvar}{the unquoted name of the column containing the labeling variable in the tipinfo dataframe. Defaults to \code{taxon}}

\item{title}{the text string to use as the title of the phylomorphospace plot}

\item{xlab}{the text label for the x axis}

\item{ylab}{the text label for the y axis}

\item{repel}{boolean indicating whether or not to use the \code{ggrepel} package to prevent overplotting of labels}

\item{edge.width}{size parameter for edges drawn on phylomorphospace. Default is 1}

\item{fontface}{font face for labels. Default is "italic".}

\item{tree.alpha}{alpha parameter for the main tree. default is 0.7}
}
\value{
the ggplot object representing the phylomorphospace
}
\description{
Make a phylomorphospace plot
}
