% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggphylomorpho.R
\name{ggphylomorpho}
\alias{ggphylomorpho}
\title{Make the phylomorphospace}
\usage{
ggphylomorpho(tree, tipinfo, xvar = PC1, yvar = PC2, factorvar = group,
  labelvar = taxon, title = "Phylomorphospace", xlab = "PC1",
  ylab = "PC2")
}
\arguments{
\item{tree}{object of class \code{phylo}}

\item{tipinfo}{a dataframe containing two continuous variables (typically PC1 and PC2 from a PCA) as well as a grouping variable and a variable of labels for each point on the plot}

\item{xvar}{the unquoted name of the column containing the x variable in the tipinfo dataframe. Defaults to \code{PC1}}

\item{yvar}{the unquoted name of the column containing the y variable in the tipinfo dataframe. Defaults to \code{PC2}}

\item{factorvar}{the unquoted name of the column containing the grouping variable in the tipinfo dataframe. Defaults to \code{group}}

\item{labelvar}{the unquoted name of the column containing the labeling variable in the tipinfo dataframe. Defaults to \code{taxon}}

\item{title}{the text string to use as the title of the phylomorphospace plot}

\item{xlab}{the text label for the x axis}

\item{ylab}{the text label for the y axis}
}
\value{
the ggplot object representing the phylomorphospace
}
\description{
Make the phylomorphospace
}
